refactor: modelからapiモジュールに移動できそうなものを移動した
