+++ modules/api/src/main/java/net/pantasystem/milktea/api/misskey/MisskeyAPI.kt
@@ -20,10 +20,12 @@ import net.pantasystem.milktea.api.misskey.groups.TransferGroupDTO
 import net.pantasystem.milktea.api.misskey.groups.UpdateGroupDTO
 import net.pantasystem.milktea.api.misskey.hashtag.SearchHashtagRequest
 import net.pantasystem.milktea.api.misskey.instance.MetaNetworkDTO
+import net.pantasystem.milktea.api.misskey.instance.RequestMeta
 import net.pantasystem.milktea.api.misskey.list.*
 import net.pantasystem.milktea.api.misskey.messaging.MessageAction
 import net.pantasystem.milktea.api.misskey.messaging.MessageDTO
 import net.pantasystem.milktea.api.misskey.messaging.RequestMessage
+import net.pantasystem.milktea.api.misskey.messaging.RequestMessageHistory
 import net.pantasystem.milktea.api.misskey.notes.*
 import net.pantasystem.milktea.api.misskey.notes.favorite.CreateFavorite
 import net.pantasystem.milktea.api.misskey.notes.favorite.DeleteFavorite
@@ -36,6 +38,7 @@ import net.pantasystem.milktea.api.misskey.notification.NotificationDTO
 import net.pantasystem.milktea.api.misskey.notification.NotificationRequest
 import net.pantasystem.milktea.api.misskey.online.user.OnlineUserCount
 import net.pantasystem.milktea.api.misskey.register.Subscription
+import net.pantasystem.milktea.api.misskey.register.SubscriptionStateNetworkDTO
 import net.pantasystem.milktea.api.misskey.register.UnSubscription
 import net.pantasystem.milktea.api.misskey.register.WebClientBaseRequest
 import net.pantasystem.milktea.api.misskey.register.WebClientRegistries
@@ -70,10 +73,7 @@ import net.pantasystem.milktea.api.misskey.v12_75_0.Show
 import net.pantasystem.milktea.api.misskey.v12_75_0.UnLike
 import net.pantasystem.milktea.api.misskey.v12_75_0.Update
 import net.pantasystem.milktea.api.misskey.v13.EmojisResponse
-import net.pantasystem.milktea.model.instance.RequestMeta
-import net.pantasystem.milktea.model.messaging.RequestMessageHistory
-import net.pantasystem.milktea.model.notes.poll.Vote
-import net.pantasystem.milktea.model.sw.register.SubscriptionState
+import net.pantasystem.milktea.api.misskey.notes.Vote
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.Headers
@@ -279,7 +279,7 @@ interface MisskeyAPI {
 
 
     @POST("api/sw/register")
-    suspend fun swRegister(@Body subscription: Subscription) : Response<SubscriptionState>
+    suspend fun swRegister(@Body subscription: Subscription) : Response<SubscriptionStateNetworkDTO>
 
     @POST("api/sw/unregister")
     suspend fun swUnRegister(@Body unSub: UnSubscription) : Response<Unit>
