+++ modules/model/src/main/java/net/pantasystem/milktea/model/sw/register/SubscriptionState.kt
@@ -3,6 +3,7 @@ package net.pantasystem.milktea.model.sw.register
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+
 @Serializable
 data class SubscriptionState(
     @SerialName("state") val state: String,
